+++ src/app/main/_components/forecastInfo/hourlyInfo.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Chart from '../chart';
 import useHourlyInfo from './useHourlyInfo';
 
@@ -8,9 +10,9 @@ export default function HourlyInfo() {
 
   return (
     <div className="p-5">
-      <div className="min-h-[180px] h-full p-4 rounded-lg bg-gray-100/70">
+      {/* <div className="min-h-[180px] h-full p-4 rounded-lg bg-gray-100/70">
         <Chart />
-      </div>
+      </div> */}
     </div>
   );
 }
